requirements.txt
absl-py==0.7.1
alabaster==0.7.12
ambient==0.1.4
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.2
AnyQt==0.0.10
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
astropy==3.1.2
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
bitarray==0.9.3
bkcharts==0.2
blaze==0.11.3
bleach==3.1.0
bokeh==1.2.0
boto==2.49.0
Bottleneck==1.2.1
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.0.0
clyent==1.2.2
colorama==0.4.1
conda==4.6.14
conda-build==3.17.8
conda-package-handling==0+unknown
conda-verify==3.3.0
contextlib2==0.5.5
cryptography==2.6.1
cycler==0.10.0
Cython==0.29.8
cytoolz==0.9.0.1
dask==1.2.2
datashape==0.5.4
decorator==4.4.0
defusedxml==0.6.0
distributed==1.28.1
Django==2.2.4
dlib==19.17.0
docutils==0.14
entrypoints==0.3
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.10
Flask==1.0.3
Flask-Cors==3.0.7
future==0.17.1
gast==0.2.2
gevent==1.4.0
glob2==0.6
gmpy2==2.0.8
google==2.0.2
greenlet==0.4.15
grpcio==1.21.1
h5py==2.9.0
heapdict==1.0.0
html5lib==1.0.1
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==0.0.0
ipykernel==5.1.0
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.19
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.13.3
jeepney==0.4
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyter-http-over-ws==0.0.6
jupyterlab==0.35.5
jupyterlab-launcher==0.13.1
jupyterlab-server==0.2.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==18.0.0
keyrings.alt==3.1.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
libarchive-c==2.8
lief==0.9.0
llvmlite==0.28.0
locket==0.2.0
lxml==4.3.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
mock==3.0.5
more-itertools==7.0.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.1
nose==1.3.7
notebook==5.7.8
numba==0.43.1
numexpr==2.6.9
numpy==1.16.4
numpydoc==0.9.1
odo==0.5.1
olefile==0.46
opencv-contrib-python==4.1.0.25
opencv-python==4.1.0.25
openpyxl==2.6.2
openTSNE==0.3.8
Orange3==3.22.0
Orange3-Associate==1.1.5
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
partd==0.3.10
path.py==12.0.1
pathlib2==2.3.3
patsy==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.4.1
pkginfo==1.5.0.1
pluggy==0.11.0
ply==3.11
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.8.0
psutil==5.6.2
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pyflakes==2.1.1
Pygments==2.4.0
pylint==2.3.1
pynndescent==0.3.0
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyqtgraph==0.10.0
pyrsistent==0.14.11
PySocks==1.7.0
pytest==4.5.0
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.0
python-louvain==0.13
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.5.1
QtPy==1.7.1
requests==2.21.0
rope==0.14.0
ruamel-yaml==0.15.46
scikit-image==0.15.0
scikit-learn==0.21.1
scipy==1.2.1
seaborn==0.9.0
SecretStorage==3.1.1
Send2Trash==1.5.0
serverfiles==0.3.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==1.2.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
Sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.0
spyder==3.3.4
spyder-kernels==0.4.4
SQLAlchemy==1.3.3
sqlparse==0.3.0
statsmodels==0.9.0
sympy==1.4
tables==3.5.1
tblib==1.4.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toolz==0.9.0
torch==1.2.0
torch-vision==0.1.6.dev0
torchvision==0.4.0
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.2
typing==3.6.4
unicodecsv==0.14.1
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
wrapt==1.11.1
wurlitzer==1.0.2
xlrd==1.2.0
XlsxWriter==1.1.8
xlwt==1.3.0
zict==0.1.4
zipp==0.3.3
